requirements.txt
pillow~=10.2.0
pip~=21.2.4
MarkupSafe~=2.1.5
werkzeug~=3.0.1
click~=8.1.7
Jinja2~=3.1.3
blinker~=1.7.0
itsdangerous~=2.1.2
numpy~=1.26.4
setuptools~=58.0.4
scipy~=1.12.0
zipp~=3.17.0
flask~=3.0.2